Project restructure Change default font size (Not tested on Windows) Fix in DisAsmTableView code (index out of range) Linux build script added

+++ gui/cpustatetableview.cpp
@@ -54,7 +54,7 @@ QModelIndex CPUStateTableView::parent(const QModelIndex &index) const {
 }
 
 int CPUStateTableView::rowCount(const QModelIndex &parent) const {
-    return _rows.size();
+    return (int) _rows.size();
 }
 
 void CPUStateTableView::addObject(std::pair<std::string, uint> value) {
